Update schema to be compatible with spark

+++ automation/DataAggregator/parquet_schema.py
@@ -4,18 +4,18 @@ PQ_SCHEMAS = dict()
 
 # site_visits
 fields = [
-    pa.field('visit_id', pa.uint64(), nullable=False),
-    pa.field('crawl_id', pa.uint32(), nullable=False),
-    pa.field('instance_id', pa.uint32(), nullable=False),
+    pa.field('visit_id', pa.int64(), nullable=False),
+    pa.field('crawl_id', pa.int32(), nullable=False),
+    pa.field('instance_id', pa.int32(), nullable=False),
     pa.field('site_url', pa.string(), nullable=False)
 ]
 PQ_SCHEMAS['site_visits'] = pa.schema(fields)
 
 # flash_cookies
 fields = [
-    pa.field('visit_id', pa.uint64(), nullable=False),
-    pa.field('crawl_id', pa.uint32(), nullable=False),
-    pa.field('instance_id', pa.uint32(), nullable=False),
+    pa.field('visit_id', pa.int64(), nullable=False),
+    pa.field('crawl_id', pa.int32(), nullable=False),
+    pa.field('instance_id', pa.int32(), nullable=False),
     pa.field('domain', pa.string()),
     pa.field('filename', pa.string()),
     pa.field('local_path', pa.string()),
@@ -27,9 +27,9 @@ PQ_SCHEMAS['flash_cookies'] = pa.schema(fields)
 
 # profile_cookies
 fields = [
-    pa.field('visit_id', pa.uint64(), nullable=False),
-    pa.field('crawl_id', pa.uint32(), nullable=False),
-    pa.field('instance_id', pa.uint32(), nullable=False),
+    pa.field('visit_id', pa.int64(), nullable=False),
+    pa.field('crawl_id', pa.int32(), nullable=False),
+    pa.field('instance_id', pa.int32(), nullable=False),
     pa.field('baseDomain', pa.string()),
     pa.field('name', pa.string()),
     pa.field('value', pa.string()),
@@ -45,9 +45,9 @@ PQ_SCHEMAS['profile_cookies'] = pa.schema(fields)
 
 # crawl_history
 fields = [
-    pa.field('visit_id', pa.uint64(), nullable=False),
-    pa.field('crawl_id', pa.uint32(), nullable=False),
-    pa.field('instance_id', pa.uint32(), nullable=False),
+    pa.field('visit_id', pa.int64(), nullable=False),
+    pa.field('crawl_id', pa.int32(), nullable=False),
+    pa.field('instance_id', pa.int32(), nullable=False),
     pa.field('command', pa.string()),
     pa.field('arguments', pa.string()),
     pa.field('bool_success', pa.int8()),
@@ -56,9 +56,9 @@ PQ_SCHEMAS['crawl_history'] = pa.schema(fields)
 
 # http_requests
 fields = [
-    pa.field('visit_id', pa.uint64(), nullable=False),
-    pa.field('crawl_id', pa.uint32(), nullable=False),
-    pa.field('instance_id', pa.uint32(), nullable=False),
+    pa.field('visit_id', pa.int64(), nullable=False),
+    pa.field('crawl_id', pa.int32(), nullable=False),
+    pa.field('instance_id', pa.int32(), nullable=False),
     pa.field('url', pa.string(), nullable=False),
     pa.field('top_level_url', pa.string()),
     pa.field('method', pa.string(), nullable=False),
@@ -74,16 +74,16 @@ fields = [
     pa.field('loading_origin', pa.string()),
     pa.field('loading_href', pa.string()),
     pa.field('req_call_stack', pa.string()),
-    pa.field('content_policy_type', pa.uint8(), nullable=False),
+    pa.field('content_policy_type', pa.int16(), nullable=False),
     pa.field('post_body', pa.string())
 ]
 PQ_SCHEMAS['http_requests'] = pa.schema(fields)
 
 # http_responses
 fields = [
-    pa.field('visit_id', pa.uint64(), nullable=False),
-    pa.field('crawl_id', pa.uint32(), nullable=False),
-    pa.field('instance_id', pa.uint32(), nullable=False),
+    pa.field('visit_id', pa.int64(), nullable=False),
+    pa.field('crawl_id', pa.int32(), nullable=False),
+    pa.field('instance_id', pa.int32(), nullable=False),
     pa.field('url', pa.string(), nullable=False),
     pa.field('method', pa.string(), nullable=False),
     pa.field('referrer', pa.string(), nullable=False),
@@ -100,9 +100,9 @@ PQ_SCHEMAS['http_responses'] = pa.schema(fields)
 
 # http_redirects
 fields = [
-    pa.field('visit_id', pa.uint64(), nullable=False),
-    pa.field('crawl_id', pa.uint32(), nullable=False),
-    pa.field('instance_id', pa.uint32(), nullable=False),
+    pa.field('visit_id', pa.int64(), nullable=False),
+    pa.field('crawl_id', pa.int32(), nullable=False),
+    pa.field('instance_id', pa.int32(), nullable=False),
     pa.field('old_channel_id', pa.string()),
     pa.field('new_channel_id', pa.string()),
     pa.field('is_temporary', pa.bool_(), nullable=False),
@@ -115,9 +115,9 @@ PQ_SCHEMAS['http_redirects'] = pa.schema(fields)
 
 # javascript
 fields = [
-    pa.field('visit_id', pa.uint64(), nullable=False),
-    pa.field('crawl_id', pa.uint32(), nullable=False),
-    pa.field('instance_id', pa.uint32(), nullable=False),
+    pa.field('visit_id', pa.int64(), nullable=False),
+    pa.field('crawl_id', pa.int32(), nullable=False),
+    pa.field('instance_id', pa.int32(), nullable=False),
     pa.field('script_url', pa.string()),
     pa.field('script_line', pa.string()),
     pa.field('script_col', pa.string()),
@@ -136,9 +136,9 @@ PQ_SCHEMAS['javascript'] = pa.schema(fields)
 
 # javascript_cookies
 fields = [
-    pa.field('visit_id', pa.uint64(), nullable=False),
-    pa.field('crawl_id', pa.uint32(), nullable=False),
-    pa.field('instance_id', pa.uint32(), nullable=False),
+    pa.field('visit_id', pa.int64(), nullable=False),
+    pa.field('crawl_id', pa.int32(), nullable=False),
+    pa.field('instance_id', pa.int32(), nullable=False),
     pa.field('change', pa.string()),
     pa.field('creationTime', pa.string()),
     pa.field('expiry', pa.string()),
@@ -146,7 +146,7 @@ fields = [
     pa.field('is_session', pa.bool_()),
     pa.field('last_accessed', pa.string()),
     pa.field('raw_host', pa.string()),
-    pa.field('expires', pa.uint64()),
+    pa.field('expires', pa.int64()),
     pa.field('host', pa.string()),
     pa.field('is_domain', pa.bool_()),
     pa.field('is_secure', pa.bool_()),
